requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anyio=4.2.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
attrs=23.1.0=pypi_0
babel=2.14.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
blas=1.0=mkl
bleach=6.1.0=pypi_0
brotli-python=1.0.9=py39hd77b12b_7
ca-certificates=2023.12.12=haa95532_0
certifi=2023.11.17=py39haa95532_0
cffi=1.16.0=py39h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pypi_0
comm=0.2.0=pypi_0
contourpy=1.2.0=pypi_0
cryptography=41.0.3=py39h3438e0d_0
cuda-cccl=12.3.101=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.3.101=0
cuda-runtime=11.8.0=0
cycler=0.12.1=pypi_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pypi_0
fastjsonschema=2.19.0=pypi_0
filelock=3.13.1=py39haa95532_0
fonttools=4.47.0=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=ha860e81_0
giflib=5.2.1=h8cc25b3_3
gmpy2=2.1.2=py39h7f96b67_0
idna=3.4=py39haa95532_0
importlib-metadata=7.0.1=pypi_0
importlib-resources=6.1.1=pypi_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.28.0=pypi_0
ipython=8.18.1=pypi_0
ipywidgets=8.1.1=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.2=py39haa95532_0
joblib=1.3.2=pypi_0
jpeg=9e=h2bbff1b_1
json5=0.9.14=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.20.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.6.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.6.0=pypi_0
jupyter-events=0.9.0=pypi_0
jupyter-lsp=2.2.1=pypi_0
jupyter-server=2.12.1=pypi_0
jupyter-server-terminals=0.5.1=pypi_0
jupyterlab=4.0.9=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.25.2=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
kiwisolver=1.4.5=pypi_0
lerc=3.0=hd77b12b_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.4.101=0
libcurand-dev=10.3.4.101=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h2bbff1b_1
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.1=py39h2bbff1b_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=3.0.2=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.8=py39h2bbff1b_0
mkl_random=1.2.4=py39h59b6b97_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py39haa95532_0
nbclient=0.9.0=pypi_0
nbconvert=7.13.1=pypi_0
nbformat=5.9.2=pypi_0
nest-asyncio=1.5.8=pypi_0
networkx=3.1=py39haa95532_0
notebook=7.0.6=pypi_0
notebook-shim=0.2.3=pypi_0
numpy=1.26.2=py39h055cbcc_0
numpy-base=1.26.2=py39h65a83cf_0
openjpeg=2.4.0=h4fc8c34_0
openpyxl=3.1.2=pypi_0
openssl=1.1.1w=h2bbff1b_0
overrides=7.4.0=pypi_0
packaging=23.2=pypi_0
pandas=2.1.4=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pillow=10.0.1=py39h045eedc_0
pip=23.3.1=py39haa95532_0
platformdirs=4.1.0=pypi_0
prometheus-client=0.19.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
psutil=5.9.7=pypi_0
pure-eval=0.2.2=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.17.2=pypi_0
pyopenssl=23.2.0=py39haa95532_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=py39haa95532_0
python=3.9.0=h6244533_2
python-dateutil=2.8.2=pypi_0
python-json-logger=2.0.7=pypi_0
pytorch=2.1.2=py3.9_cuda11.8_cudnn8_0
pytorch-cuda=11.8=h24eeafa_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pywin32=306=pypi_0
pywinpty=2.0.12=pypi_0
pyyaml=6.0.1=py39h2bbff1b_0
pyzmq=25.1.2=pypi_0
qtconsole=5.5.1=pypi_0
qtpy=2.4.1=pypi_0
referencing=0.32.0=pypi_0
requests=2.31.0=py39haa95532_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.15.2=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.4=pypi_0
seaborn=0.13.0=pypi_0
send2trash=1.8.2=pypi_0
setuptools=68.2.2=py39haa95532_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h2bbff1b_0
stack-data=0.6.3=pypi_0
sympy=1.12=py39haa95532_0
tbb=2021.8.0=h59b6b97_0
terminado=0.18.0=pypi_0
threadpoolctl=3.2.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h2bbff1b_0
tomli=2.0.1=pypi_0
torchaudio=2.1.2=pypi_0
torchvision=0.16.2=pypi_0
tornado=6.4=pypi_0
traitlets=5.14.0=pypi_0
types-python-dateutil=2.8.19.14=pypi_0
typing_extensions=4.7.1=py39haa95532_0
tzdata=2023.3=pypi_0
uri-template=1.3.0=pypi_0
urllib3=1.26.18=py39haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.12=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.7.0=pypi_0
wheel=0.41.2=py39haa95532_0
widgetsnbextension=4.0.9=pypi_0
win_inet_pton=1.1.0=py39haa95532_0
xz=5.4.5=h8cc25b3_0
yaml=0.2.5=he774522_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
